Refactor shared runtime contracts into webcodex-core
