+++ crates/webcodex-core/Cargo.toml
@@ -7,6 +7,7 @@ edition.workspace = true
 [dependencies]
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+sha2 = "0.10"
 tokio = { version = "1", features = ["io-util"] }
 
 [dev-dependencies]
