+++ src/route_metadata/consoles.rs
@@ -3,7 +3,7 @@ use super::{
     route, AuditClass::*, OpenApiVisibility::*, RouteAuth, RouteId::*, RouteMethod::*, RouteSpec,
     RouteSurface::*,
 };
-use crate::auth::scopes::{
+use webcodex_core::authority::{
     OAuthRouteScopePolicy::*, SCOPE_COMMUNICATION_MANAGE, SCOPE_COMMUNICATION_READ,
     SCOPE_PROJECT_READ, SCOPE_RUNTIME_READ, SCOPE_SESSION_COLLABORATE,
 };
