+++ src/tool_runtime/tool_definition/communication.rs
@@ -3,7 +3,6 @@ use super::{
     def, model_spec, permission_risk, require_all_scopes, ToolDefinition, PERMISSION_RISK_WRITE,
     TOOL_CATEGORY_COMMUNICATION,
 };
-use crate::auth::scopes::{COMMUNICATION_MANAGE_SCOPES, COMMUNICATION_READ_SCOPES};
 use crate::tool_runtime::metadata::{
     ToolPathHint::None as NoPath,
     ToolRisk::{CommunicationManage, Read},
@@ -18,6 +17,7 @@ use crate::tool_runtime::registry::input_schemas::{
     post_conversation_message_input_schema, read_conversation_input_schema,
     update_agent_identity_input_schema,
 };
+use webcodex_core::authority::{COMMUNICATION_MANAGE_SCOPES, COMMUNICATION_READ_SCOPES};
 
 pub(super) const DEFINITIONS: &[ToolDefinition] = &[
     require_all_scopes(
